Bug 2013273 - Make nsMencloseNotation an enum class. r=layout-reviewers,dshin
Differential Revision: https://phabricator.services.mozilla.com/D280936

UltraBlame original commit: fdd3d89ebcd4389da60c5885796689806d3b53c7

+++ layout/mathml/nsMathMLFrame.cpp
@@ -282,7 +282,7 @@ void nsDisplayMathMLBar::Paint(nsDisplayListBuilder* aBuilder,
 void nsMathMLFrame::DisplayBar(nsDisplayListBuilder* aBuilder, nsIFrame* aFrame,
                                const nsRect& aRect,
                                const nsDisplayListSet& aLists,
-                               uint32_t aIndex) {
+                               uint16_t aIndex) {
   if (!aFrame->StyleVisibility()->IsVisible() || aRect.IsEmpty()) {
     return;
   }
